Added changes to Repository layer and added crud operations to CategoryController.

+++ ApplicationService/DTOs/ProductDTO.cs
@@ -8,14 +8,13 @@ namespace ApplicationService.DTOs
 {
     public class ProductDTO
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         [Required]
         public string ProductName { get; set; }
         public string Description { get; set; }
         public DateTime? Release { get; set; }
         public double Price { get; set; }
         public string Image { get; set; }
-
         public virtual CategoryDTO Category { get; set; }
     }
 }
