+++ Data/Entitites/Product.cs
@@ -9,12 +9,13 @@ namespace Data.Entitites
     {
         [Required]
         public string ProductName { get; set; }
+
         public string Description { get; set; }
         public DateTime? Release { get; set; }
         public double Price { get; set; }
         public string Image { get; set; }
 
-        public long CategoryId { get; set; }
+        public int CategoryId { get; set; }
         public virtual Category Category { get; set; }
     }
 }
